+++ ktrain/text/textutils.py
@@ -203,8 +203,12 @@ def detect_lang(texts, sample_size=32):
         except:
             continue
     if len(lst) == 0: 
-        raise Exception('could not detect language in random sample of %s docs. Are you sure you provided a list of strings?'  % (sample_size))
-    return max(set(lst), key=lst.count)
+        warnings.warn('Defaulting to English: could not detect language in random sample of %s docs. Are you sure you provided a list of strings?'  % (sample_size))
+        lang = 'en'
+    else:
+        lang = max(set(lst), key=lst.count)
+    #return max(set(lst), key=lst.count)
+    return lang